2.7.3：将固定位数的整型位数与之前保持一致

+++ src/Algorithms/Match.h
@@ -23,7 +23,7 @@
 
 namespace CZ
 {
-    using MatchRank = uint64_t;
+    using MatchRank = uint32_t;
 
     enum MatchMethod
     {
